keep DepType in ckb-types.
